requirements-travis.txt
nose==1.3.7
numpy==1.13.0
six==1.10.0
ruamel.yaml==0.15.6
msgpack-python==0.4.8
coverage==4.4.1
coveralls==1.1
pymongo==3.4.0